Add template argument to classes CompletePrediction and PartialPrediction.

+++ cpp/subprojects/seco/include/mlrl/seco/data/matrix_coverage_dense.hpp
@@ -52,8 +52,8 @@ namespace seco {
              */
             void increaseCoverage(uint32 row, View<uint32>::const_iterator majorityLabelIndicesBegin,
                                   View<uint32>::const_iterator majorityLabelIndicesEnd,
-                                  View<float64>::const_iterator predictionBegin,
-                                  View<float64>::const_iterator predictionEnd,
+                                  View<float32>::const_iterator predictionBegin,
+                                  View<float32>::const_iterator predictionEnd,
                                   CompleteIndexVector::const_iterator indicesBegin,
                                   CompleteIndexVector::const_iterator indicesEnd);
 
@@ -73,8 +73,8 @@ namespace seco {
              */
             void increaseCoverage(uint32 row, View<uint32>::const_iterator majorityLabelIndicesBegin,
                                   View<uint32>::const_iterator majorityLabelIndicesEnd,
-                                  View<float64>::const_iterator predictionBegin,
-                                  View<float64>::const_iterator predictionEnd,
+                                  View<float32>::const_iterator predictionBegin,
+                                  View<float32>::const_iterator predictionEnd,
                                   PartialIndexVector::const_iterator indicesBegin,
                                   PartialIndexVector::const_iterator indicesEnd);
 
@@ -94,8 +94,8 @@ namespace seco {
              */
             void decreaseCoverage(uint32 row, View<uint32>::const_iterator majorityLabelIndicesBegin,
                                   View<uint32>::const_iterator majorityLabelIndicesEnd,
-                                  View<float64>::const_iterator predictionBegin,
-                                  View<float64>::const_iterator predictionEnd,
+                                  View<float32>::const_iterator predictionBegin,
+                                  View<float32>::const_iterator predictionEnd,
                                   CompleteIndexVector::const_iterator indicesBegin,
                                   CompleteIndexVector::const_iterator indicesEnd);
 
@@ -115,8 +115,8 @@ namespace seco {
              */
             void decreaseCoverage(uint32 row, View<uint32>::const_iterator majorityLabelIndicesBegin,
                                   View<uint32>::const_iterator majorityLabelIndicesEnd,
-                                  View<float64>::const_iterator predictionBegin,
-                                  View<float64>::const_iterator predictionEnd,
+                                  View<float32>::const_iterator predictionBegin,
+                                  View<float32>::const_iterator predictionEnd,
                                   PartialIndexVector::const_iterator indicesBegin,
                                   PartialIndexVector::const_iterator indicesEnd);
     };
